+++ routes/channels.php
@@ -11,6 +11,5 @@
 |
 */
 
-Broadcast::channel('App.User.{id}', function ($user, $id) {
-    return (int) $user->id === (int) $id;
-});
+Broadcast::channel('orders.{mrn}');
+Broadcast::channel('labs.{mrn}');