+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 This project adheres to [semantic versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Changing the patient for orders or labs is broadcast as an event and
+  shown to users without needing refreshing
+- Redis queue for event broadcasting
 
-## [0.7.2] - 2019-01-31
+## [0.7.3] - 2019-01-31
 ### Fixed
 - The signature submit form is shown again after scanning medication
 - Patient MARs are updated to provide feedback when signatures are submitted